Adjust source model assignment and implement pinned files handling
The changes introduced improve how the source model for `NotesFolderModel` is assigned in `MainWindow.cpp`. This fixes the issue of creating a list of pinned files, which is now done after the directory is loaded in the source `QFileSystemModel`. Additionally, handling for pinned files was implemented, where files are sorted based on whether they're pinned and the sort order value.